# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Service Ticket Management System API"
  description: "API to manage a Mechanic Service Ticket System"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# Security Definitions
# =========================
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# =========================
# Paths
# =========================
paths: 
...
  /customers/login: #Customers Login
    post:
      tags:
        - Customers
      summary: Customer Login (generate token)
      description: This endpoint allows a customer to log in.
      parameters:
        - in: body
          name: body
          description: Login credentials
          required: true
          schema:
            $ref: '#/definitions/LoginPayload'
      responses:
        200:
          description: Login successful
          schema:
            $ref: '#/definitions/LoginResponse'
          examples:
            application/json:
              {
                "status": "success",
                "message": "Successfully logged in.",
                "token": "eyJhbGciOiJIUz....."
              }
        401:
          description: Invalid credentials
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Invalid email or password.
              status: error

  # --------- Customer CRUD ---------
  /customers: #Customer CRUD
    post: #Create
      tags:
        - Customers
      summary: Create a new customer
      description: This endpoint allows you to create a new customer.
      parameters:
        - in: body
          name: body
          description: Customer object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/CustomerPayload'
      responses:
        201:
          description: Customer created successfully
          schema:
            $ref: '#/definitions/CustomerResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "John Doe",
                "phone": "0987654321",
                "email": "jd@app.com"
              }
        400:
          description: Email already exists
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Email already exists.
              status: error          
    get: #Read
      tags:
        - Customers
      summary: Get customers (with pagination)
      description: This endpoint returns a list of customers with pagination. 
        **You can use the `page` and `per_page` query parameters to control pagination.**
      parameters:
        - in: query
          name: page
          type: integer
          description: Page number for pagination.
        - in: query
          name: per_page
          type: integer
          description: Number of customers per page.
      responses:
        200:
          description: Return all customers
          schema:
            $ref: '#/definitions/CustomerListResponse'
            examples:
              application/json:
                [
                  {
                    "id": 1,
                    "name": "John Doe",
                    "phone": "0987654321",
                    "email": "jd@app.com",
                  }
                ]    
    put: #Update
      tags:
        - Customers
      summary: Update a customer (Token required)
      description: Update an existing customer's details.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Information used to update a customer
          required: true
          schema:
            $ref: '#/definitions/CustomerPayload'
      responses:
        200:
          description: Customer updated successfully
          schema:
            $ref: '#/definitions/CustomerResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "John Doe",
                "phone": "0987654321",
                "email": "jd@app.com",
              }
        400:
          description: Email already exists
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Email already exists.
              status: error
    delete: #Delete
      tags:
        - Customers
      summary: Delete a customer (Token required)
      description: This endpoint allows you to delete a customer.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Customer deleted successfully
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Successfully deleted customer 1.
              status: success

  # --------- Customers Get By ID ---------
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: Get customer by ID
      description: Retrieve a customer's details by their unique ID.
      parameters:
        - in: path
          name: customer_id
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
          description: The ID of the customer to retrieve.
      responses:
        200:
          description: Customer found
          schema:
            $ref: '#/definitions/CustomerResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "John Doe",
                "phone": "0987654321",
                "email": "jd@app.com",
              }
        404:
          description: Customer not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Customer not found!
              status: error

  # --------- Customers Search ---------              
  /customers/search:
    get:
      tags:
        - Customers
      summary: Search customers
      description: This endpoint allows you to search for customers by name and/or email.
        **At least one of `name` or `email` must be provided as a query parameter.**.
      parameters:
        - in: query
          name: name
          type: string
          description: Search term to filter customers by name.
        - in: query
          name: email
          type: string
          description: Search term to filter customers by email.
      responses:
        200:
          description: Return customers matching the search term
          schema:
            $ref: '#/definitions/CustomerListResponse'
            examples:
              application/json:
                [
                  {
                    "id": 1,
                    "name": "John Doe",
                    "phone": "0987654321",
                    "email": "jd@app.com",
                  }
                ]
        400:
          description: No query parameters provided
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: At least one search parameter (name or email) is required.
              status: error
        404:
          description: No customers found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: No customers found.
              status: error

  # --------- Customers Ticket ---------        
  /customers/my-tickets:
    get:
      tags:
        - Customers
      summary: Get my tickets (Token required)
      description: This endpoint allows you to get all tickets of the customer.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Return all tickets of the customer
          schema:
            $ref: '#/definitions/CustomerTicketsResponse'
          examples:
            application/json:
              {
                "customer": {
                  "email": "jd@app.com",
                  "id": 1,
                  "name": "John Doe",
                  "phone": "123-456-789"
                },
                "id": 1,
                "service_tickets": []
              }
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Token is missing.

  # --------- Parts Description Endpoints ---------         
...
definitions: 
...
  LoginPayload:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  LoginResponse:
    type: object
    properties:
      status:
        type: string
      message:
        type: string
      token:
        type: string
  
  # --------- Message Schemas ---------
  MessageSchemaResponse:
    type: object
    properties:
      message:
        type: string
      status:
        type: string

  # --------- Mechanic Schemas ---------
...
  CustomerPayload:
    type: object
    properties:
      name:
        type: string
      phone:
        type: string
      email:
        type: string
      password:
        type: string
  CustomerResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      email:
        type: string
  CustomerListResponse:
    type: array
    items:
      $ref: '#/definitions/CustomerResponse'
  CustomerTicketsResponse:
    type: object
    properties:
      customer:
        $ref: '#/definitions/CustomerResponse'
      id:
        type: integer
      service_tickets:
        type: array

#--------------- Part Description Schemas-------
...